+++ src/mako/application/services/LoggerService.php
@@ -40,7 +40,7 @@ class LoggerService extends Service
 			{
 				$user = ['ip_address' => $this->container->get(Request::class)->getIp()];
 
-				if($this->container->has(Gatekeeper::class) && ($gatekeeperUser = $this->container->get(Gatekeeper::class)->getUser()) !== null)
+				if($this->container->hasInstanceOf(Gatekeeper::class) && ($gatekeeperUser = $this->container->get(Gatekeeper::class)->getUser()) !== null)
 				{
 					$user += ['id' => $gatekeeperUser->getId(), 'username' => $gatekeeperUser->getUsername()];
 				}